refactor: 모임 댓글 신고 내역 조회 API 수정 및 코드 리팩토링
Swagger 작성 및 모임 댓글 신고 누적 개수 쿼리 추가

+++ suldaksuldak-core/src/main/java/com/sulsul/suldaksuldak/dto/party/comment/PartyCommentDto.java
@@ -15,11 +15,10 @@ public class PartyCommentDto {
     String partyName;
     String partyFileNm;
     String groupComment;
-//    Long commentCnt;
     Integer commentDep;
     Boolean isDelete;
     Boolean isModified;
-    Integer warningCnt;
+    Long warningCnt;
     LocalDateTime createdAt;
     LocalDateTime modifiedAt;
 }
